Primer caso de uso (a medias).
