--- src/main/java/edu/fiuba/algo3/modelo/Equipamiento/Armadura.java
@@ -1,7 +1,0 @@
-package edu.fiuba.algo3.modelo.Equipamiento;
-
-public class Armadura implements Equipamiento{
-    public Equipamiento mejorarEquipamiento() {
-        return new EscudoYEspada();
-    }
-}
