--- src/main/java/edu/fiuba/algo3/modelo/Equipamiento/Casco.java
@@ -1,7 +1,0 @@
-package edu.fiuba.algo3.modelo.Equipamiento;
-
-public class Casco implements Equipamiento{
-    public Equipamiento mejorarEquipamiento() {
-        return new Armadura();
-    }
-}
