--- src/main/java/edu/fiuba/algo3/modelo/Equipamiento/Desequipado.java
@@ -1,7 +1,0 @@
-package edu.fiuba.algo3.modelo.Equipamiento;
-
-public class Desequipado implements Equipamiento{
-    public Equipamiento mejorarEquipamiento() {
-        return new Casco();
-    }
-}
