+++ src/main/java/edu/fiuba/algo3/modelo/Gladiador.java
@@ -1,12 +1,18 @@
 package edu.fiuba.algo3.modelo;
-//Energia inicial -> como constante ?
+
+import edu.fiuba.algo3.modelo.seniorities.Novato;
+import edu.fiuba.algo3.modelo.equipamientos.Desequipado;
+
 public class Gladiador {
+    //Declaramos constantes ?
+    private static final int ENERGIA_INICIAL = 20;
+
     private int energia;
     private Seniority seniority;
     private Equipamiento equipamiento;
 
     public Gladiador() {
-        this.energia = 20;
+        this.energia = ENERGIA_INICIAL;
         this.seniority = new Novato();
         this.equipamiento = new Desequipado();
     }
@@ -19,8 +25,15 @@ public class Gladiador {
         this.energia -= energia;
     }
 
-    public boolean tieneEnergia(int energia) {
-        return (this.energia == energia);
+    public boolean energiaIgualA(int energia) {
+        return this.energia == energia;
     }
 
+    public Equipamiento equipamientoGladiador(){
+        return this.equipamiento;
+    }
+
+   // public void recibirAtaque(Obstaculo unObstaculo){
+    //    this.energia -= unObstaculo.reducirEnergia(this.equipamiento);
+    //}
 }
