+++ src/main/java/edu/fiuba/algo3/modelo/Jugador.java
@@ -1,16 +1,18 @@
 package edu.fiuba.algo3.modelo;
 
-import java.util;
 
 //Decorator pattern para equipamiento ?¿
-public class Jugador {
-    private int energia;
+public class Jugador{
     private Gladiador gladiador;
 
     //private Casilla casilla;
 
     public Jugador(Gladiador gladiador) {
         this.gladiador = gladiador;
+    }
+
+    public boolean energiaIgualA(int energia){
+        return this.gladiador.energiaIgualA(energia);
     }
