--- src/main/java/edu/fiuba/algo3/modelo/Seniority/Senior.java
@@ -1,8 +1,0 @@
-package edu.fiuba.algo3.modelo.Seniority;
-
-public class Senior implements Seniority{
-    public Seniority ascender(int turno) {
-        return this;
-    }
-
-}
