+++ src/main/java/edu/fiuba/algo3/modelo/equipamientos/Armadura.java
@@ -1,0 +1,9 @@
+package edu.fiuba.algo3.modelo.equipamientos;
+
+import edu.fiuba.algo3.modelo.Equipamiento;
+
+public class Armadura implements Equipamiento {
+    public Equipamiento mejorarEquipamiento() {
+        return new EscudoYEspada();
+    }
+}
