+++ src/main/java/edu/fiuba/algo3/modelo/equipamientos/Desequipado.java
@@ -1,0 +1,10 @@
+package edu.fiuba.algo3.modelo.equipamientos;
+
+import edu.fiuba.algo3.modelo.Equipamiento;
+
+public class Desequipado implements Equipamiento {
+    public Equipamiento mejorarEquipamiento() {
+        return new Casco();
+    }
+
+}
