+++ src/main/java/edu/fiuba/algo3/modelo/equipamientos/EscudoYEspada.java
@@ -1,4 +1,6 @@
-package edu.fiuba.algo3.modelo.Equipamiento;
+package edu.fiuba.algo3.modelo.equipamientos;
+
+import edu.fiuba.algo3.modelo.Equipamiento;
 
 public class EscudoYEspada implements Equipamiento {
     public Equipamiento mejorarEquipamiento() {
