+++ src/main/java/edu/fiuba/algo3/modelo/equipamientos/Llave.java
@@ -1,4 +1,6 @@
-package edu.fiuba.algo3.modelo.Equipamiento;
+package edu.fiuba.algo3.modelo.equipamientos;
+
+import edu.fiuba.algo3.modelo.Equipamiento;
 
 public class Llave implements Equipamiento {
     public Equipamiento mejorarEquipamiento() {
