+++ src/main/java/edu/fiuba/algo3/modelo/seniorities/Senior.java
@@ -1,0 +1,10 @@
+package edu.fiuba.algo3.modelo.seniorities;
+
+import edu.fiuba.algo3.modelo.Seniority;
+
+public class Senior implements Seniority {
+    public Seniority ascender(int turno) {
+        return this;
+    }
+
+}
