--- src/main/test/edu/fiuba/algo3/entregas/MessageTest.java
@@ -1,20 +1,0 @@
-package edu.fiuba.algo3.entregas;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class MessageTest {
-    @Test
-    public void messageGreeting() {
-        Message message = new Message("Hola Mundo!", "Hello world!");
-
-        assertEquals("Hello world!", message.greet("us"));
-    }
-
-    @Test
-    public void messageGreetingDefaultLanguage() {
-        Message message = new Message("Hola Mundo!", "Hello world!");
-
-        assertEquals("Hola Mundo!", message.greet());
-    }
-}
